refactor(snapshot): move new_from + new_from_crc tests into incremental.rs
Both tests exercise `Snapshot::builder_from` and the incremental update path. They
belong with the other incremental tests rather than in `mod.rs`. Also rewrite the
incremental.rs module doc to drop the "state machine" terminology, which a reviewer
read as a reference to the builder pattern.

Addresses PR review feedback.
