+++ kernel/src/snapshot/mod.rs
@@ -1282,341 +1282,6 @@ mod tests {
         assert_eq!(snapshot.schema(), expected);
     }
 
-    // interesting cases for testing Snapshot::new_from:
-    // 1. new version < existing version
-    // 2. new version == existing version
-    // 3. new version > existing version AND
-    //   a. log segment hasn't changed
-    //   b. log segment for old..=new version has a checkpoint (with new protocol/metadata)
-    //   b. log segment for old..=new version has no checkpoint
-    //     i. commits have (new protocol, new metadata)
-    //     ii. commits have (new protocol, no metadata)
-    //     iii. commits have (no protocol, new metadata)
-    //     iv. commits have (no protocol, no metadata)
-    #[tokio::test]
-    async fn test_snapshot_new_from() -> DeltaResult<()> {
-        let path =
-            std::fs::canonicalize(PathBuf::from("./tests/data/table-with-dv-small/")).unwrap();
-        let url = url::Url::from_directory_path(path).unwrap();
-
-        let engine = SyncEngine::new();
-        let old_snapshot = Snapshot::builder_for(url.clone())
-            .at_version(1)
-            .build(&engine)
-            .unwrap();
-        // 1. new version < existing version: error
-        let snapshot_res = Snapshot::builder_from(old_snapshot.clone())
-            .at_version(0)
-            .build(&engine);
-        assert!(matches!(
-            snapshot_res,
-            Err(Error::Generic(msg)) if msg == "Requested snapshot version 0 is older than snapshot hint version 1"
-        ));
-
-        // 2. new version == existing version
-        let snapshot = Snapshot::builder_from(old_snapshot.clone())
-            .at_version(1)
-            .build(&engine)
-            .unwrap();
-        let expected = old_snapshot.clone();
-        assert_eq!(snapshot, expected);
-
-        // tests Snapshot::new_from by:
-        // 1. creating a snapshot with new API for commits 0..=2 (based on old snapshot at 0)
-        // 2. comparing with a snapshot created directly at version 2
-        //
-        // the commits tested are:
-        // - commit 0 -> base snapshot at this version
-        // - commit 1 -> final snapshots at this version
-        //
-        // in each test we will modify versions 1 and 2 to test different scenarios
-        fn test_new_from(store: Arc<InMemory>) -> DeltaResult<()> {
-            let table_root = "memory:///";
-            let engine = SyncEngine::new_with_store(store);
-            let base_snapshot = Snapshot::builder_for(table_root)
-                .at_version(0)
-                .build(&engine)?;
-            let snapshot = Snapshot::builder_from(base_snapshot.clone())
-                .at_version(1)
-                .build(&engine)?;
-            let expected = Snapshot::builder_for(table_root)
-                .at_version(1)
-                .build(&engine)?;
-            assert_eq!(snapshot, expected);
-            Ok(())
-        }
-
-        // for (3) we will just engineer custom log files
-        let store = Arc::new(InMemory::new());
-        // everything will have a starting 0 commit with commitInfo, protocol, metadata
-        let commit0 = vec![
-            json!({
-                "commitInfo": {
-                    "timestamp": 1587968586154i64,
-                    "operation": "WRITE",
-                    "operationParameters": {"mode":"ErrorIfExists","partitionBy":"[]"},
-                    "isBlindAppend":true
-                }
-            }),
-            json!({
-                "protocol": {
-                    "minReaderVersion": 1,
-                    "minWriterVersion": 2
-                }
-            }),
-            json!({
-                "metaData": {
-                    "id":"5fba94ed-9794-4965-ba6e-6ee3c0d22af9",
-                    "format": {
-                        "provider": "parquet",
-                        "options": {}
-                    },
-                    "schemaString": "{\"type\":\"struct\",\"fields\":[{\"name\":\"id\",\"type\":\"integer\",\"nullable\":true,\"metadata\":{}},{\"name\":\"val\",\"type\":\"string\",\"nullable\":true,\"metadata\":{}}]}",
-                    "partitionColumns": [],
-                    "configuration": {},
-                    "createdTime": 1587968585495i64
-                }
-            }),
-        ];
-        let table_root = "memory:///";
-        commit(table_root, store.as_ref(), 0, commit0.clone()).await;
-        // 3. new version > existing version
-        // a. no new log segment
-        let engine = SyncEngine::new_with_store(Arc::new(store.fork()));
-        let base_snapshot = Snapshot::builder_for(table_root)
-            .at_version(0)
-            .build(&engine)?;
-        let snapshot = Snapshot::builder_from(base_snapshot.clone()).build(&engine)?;
-        let expected = Snapshot::builder_for(table_root)
-            .at_version(0)
-            .build(&engine)?;
-        assert_eq!(snapshot, expected);
-        // version exceeds latest version of the table = err
-        assert!(matches!(
-            Snapshot::builder_from(base_snapshot.clone()).at_version(1).build(&engine),
-            Err(Error::Generic(msg)) if msg == "Requested snapshot version 1 is not available: no new commits were found after existing snapshot version 0"
-        ));
-
-        // b. log segment for old..=new version has a checkpoint (with new protocol/metadata)
-        let store_3a = store.fork();
-        let mut checkpoint1 = commit0.clone();
-        commit(table_root, &store_3a, 1, commit0.clone()).await;
-        checkpoint1[1] = json!({
-            "protocol": {
-                "minReaderVersion": 2,
-                "minWriterVersion": 5
-            }
-        });
-        checkpoint1[2]["partitionColumns"] = serde_json::to_value(["some_partition_column"])?;
-
-        let handler = engine.json_handler();
-        let json_strings: StringArray = checkpoint1
-            .into_iter()
-            .map(|json| json.to_string())
-            .collect::<Vec<_>>()
-            .into();
-        let parsed = handler
-            .parse_json(
-                string_array_to_engine_data(json_strings),
-                crate::actions::get_commit_schema().clone(),
-            )
-            .unwrap();
-        let checkpoint = ArrowEngineData::try_from_engine_data(parsed).unwrap();
-        let checkpoint: RecordBatch = checkpoint.into();
-
-        // Write the record batch to a Parquet file
-        let mut buffer = vec![];
-        let mut writer = ArrowWriter::try_new(&mut buffer, checkpoint.schema(), None)?;
-        writer.write(&checkpoint)?;
-        writer.close()?;
-
-        store_3a
-            .put(
-                &delta_path_for_version(1, "checkpoint.parquet"),
-                buffer.into(),
-            )
-            .await
-            .unwrap();
-        test_new_from(store_3a.into())?;
-
-        // c. log segment for old..=new version has no checkpoint
-        // i. commits have (new protocol, new metadata)
-        let store_3c_i = Arc::new(store.fork());
-        let mut commit1 = commit0.clone();
-        commit1[1] = json!({
-            "protocol": {
-                "minReaderVersion": 2,
-                "minWriterVersion": 5
-            }
-        });
-        commit1[2]["partitionColumns"] = serde_json::to_value(["some_partition_column"])?;
-        commit(table_root, store_3c_i.as_ref(), 1, commit1).await;
-        test_new_from(store_3c_i.clone())?;
-
-        // new commits AND request version > end of log
-        let engine = SyncEngine::new_with_store(store_3c_i);
-        let base_snapshot = Snapshot::builder_for(table_root)
-            .at_version(0)
-            .build(&engine)?;
-        assert!(matches!(
-            Snapshot::builder_from(base_snapshot.clone()).at_version(2).build(&engine),
-            Err(Error::Generic(msg)) if msg == "LogSegment end version 1 not the same as the specified end version 2"
-        ));
-
-        // ii. commits have (new protocol, no metadata)
-        let store_3c_ii = store.fork();
-        let mut commit1 = commit0.clone();
-        commit1[1] = json!({
-            "protocol": {
-                "minReaderVersion": 2,
-                "minWriterVersion": 5
-            }
-        });
-        commit1.remove(2); // remove metadata
-        commit(table_root, &store_3c_ii, 1, commit1).await;
-        test_new_from(store_3c_ii.into())?;
-
-        // iii. commits have (no protocol, new metadata)
-        let store_3c_iii = store.fork();
-        let mut commit1 = commit0.clone();
-        commit1[2]["partitionColumns"] = serde_json::to_value(["some_partition_column"])?;
-        commit1.remove(1); // remove protocol
-        commit(table_root, &store_3c_iii, 1, commit1).await;
-        test_new_from(store_3c_iii.into())?;
-
-        // iv. commits have (no protocol, no metadata)
-        let store_3c_iv = store.fork();
-        let commit1 = vec![commit0[0].clone()];
-        commit(table_root, &store_3c_iv, 1, commit1).await;
-        test_new_from(store_3c_iv.into())?;
-
-        Ok(())
-    }
-
-    // test new CRC in new log segment (old log segment has old CRC)
-    #[tokio::test]
-    async fn test_snapshot_new_from_crc() -> Result<(), Box<dyn std::error::Error>> {
-        let store = Arc::new(InMemory::new());
-        let table_root = "memory:///";
-        let engine = SyncEngine::new_with_store(store.clone());
-        let protocol = |reader_version, writer_version| {
-            json!({
-                "protocol": {
-                    "minReaderVersion": reader_version,
-                    "minWriterVersion": writer_version
-                }
-            })
-        };
-        let metadata = json!({
-            "metaData": {
-                "id":"5fba94ed-9794-4965-ba6e-6ee3c0d22af9",
-                "format": {
-                    "provider": "parquet",
-                    "options": {}
-                },
-                "schemaString": "{\"type\":\"struct\",\"fields\":[{\"name\":\"id\",\"type\":\"integer\",\"nullable\":true,\"metadata\":{}},{\"name\":\"val\",\"type\":\"string\",\"nullable\":true,\"metadata\":{}}]}",
-                "partitionColumns": [],
-                "configuration": {},
-                "createdTime": 1587968585495i64
-            }
-        });
-        let commit0 = vec![
-            json!({
-                "commitInfo": {
-                    "timestamp": 1587968586154i64,
-                    "operation": "WRITE",
-                    "operationParameters": {"mode":"ErrorIfExists","partitionBy":"[]"},
-                    "isBlindAppend":true
-                }
-            }),
-            protocol(1, 1),
-            metadata.clone(),
-        ];
-        let commit1 = vec![
-            json!({
-                "commitInfo": {
-                    "timestamp": 1587968586154i64,
-                    "operation": "WRITE",
-                    "operationParameters": {"mode":"ErrorIfExists","partitionBy":"[]"},
-                    "isBlindAppend":true
-                }
-            }),
-            protocol(1, 2),
-        ];
-
-        // commit 0 and 1 jsons
-        commit(table_root, &store, 0, commit0.clone()).await;
-        commit(table_root, &store, 1, commit1).await;
-
-        // Test CRC handling during incremental snapshot update (v0 -> v1).
-        // The new log listing starts at v1, so the new log segment doesn't find 0.crc.
-        // a) Only 0.crc exists: resolve_crc falls back to old segment's 0.crc.
-        // b) Both 0.crc and 1.crc exist: resolve_crc picks up 1.crc.
-        let crc = json!({
-            "table_size_bytes": 100,
-            "num_files": 1,
-            "num_metadata": 1,
-            "num_protocol": 1,
-            "metadata": metadata,
-            "protocol": protocol(1, 1),
-        });
-
-        // put the old crc
-        let path = delta_path_for_version(0, "crc");
-        store.put(&path, crc.to_string().into()).await?;
-
-        // base snapshot is at version 0
-        let base_snapshot = Snapshot::builder_for(table_root)
-            .at_version(0)
-            .build(&engine)?;
-
-        // a) only 0.crc exists -- falls back to old segment's 0.crc
-        let snapshot = Snapshot::builder_from(base_snapshot.clone())
-            .at_version(1)
-            .build(&engine)?;
-        assert_eq!(
-            snapshot
-                .log_segment
-                .listed
-                .latest_crc_file
-                .as_ref()
-                .unwrap()
-                .version,
-            0
-        );
-
-        // b) both 0.crc and 1.crc exist -- resolve_crc picks up 1.crc
-        let path = delta_path_for_version(1, "crc");
-        let crc = json!({
-            "table_size_bytes": 100,
-            "num_files": 1,
-            "num_metadata": 1,
-            "num_protocol": 1,
-            "metadata": metadata,
-            "protocol": protocol(1, 2),
-        });
-        store.put(&path, crc.to_string().into()).await?;
-        let snapshot = Snapshot::builder_from(base_snapshot.clone())
-            .at_version(1)
-            .build(&engine)?;
-        let expected = Snapshot::builder_for(table_root)
-            .at_version(1)
-            .build(&engine)?;
-        assert_eq!(snapshot, expected);
-        assert_eq!(
-            snapshot
-                .log_segment
-                .listed
-                .latest_crc_file
-                .as_ref()
-                .unwrap()
-                .version,
-            1
-        );
-
-        Ok(())
-    }
-
     #[test]
     fn test_read_table_with_missing_last_checkpoint() {
         // this table doesn't have a _last_checkpoint file
